web/requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.1
redis==3.5.3
requests==2.25.1

